Shorten gamelibs to game

+++ modlib/internal/cvar/cvar.go
@@ -20,8 +20,8 @@ func makeCVar(name, value string) engine.CVar {
 	floatVal, _ := strconv.ParseFloat(value, 32)
 	cvar := engine.RawCVar{
 		// Probably don't need to free these strings?
-		Name:   uintptr(unsafe.Pointer(C.CString(name))),
-		String: uintptr(unsafe.Pointer(C.CString(value))),
+		Name:   unsafe.Pointer(C.CString(name)),
+		String: unsafe.Pointer(C.CString(value)),
 		Value:  float32(floatVal),
 	}
 	return engine.MakeCVar(unsafe.Pointer(&cvar))
